Added new generic device identification based on device class information from the ISY SDK. When device is not found in the isydevicetypes.json file the system will attempt to use type information to identify the device generically. Unfortunately the categories are not as clean as needed, some overrides will be needed which is why isydevicetypes.json will stay for now.
Minor fixup on number rounding for dim levels. Was rounding down resulting in small off by one errors.

Minro fix where device was out of order in isydevicetypes.json.

Renamed test.js to scratch.js to better represent it's purpose. It's not a test it's a simple test app when I am doing quick testing.

Added a mocha based test suite based on working with the fake-isy-994i project. If you run an instance of that server locally this set of mocha tests will workout the API.